Patch and Delete Rest Api methods

+++ 04-spring-rest-crud/restproject/src/main/java/com/example/springboot/restproject/rest/EmployeeRestController.java
@@ -1,22 +1,26 @@
 package com.example.springboot.restproject.rest;
 
-import com.example.springboot.restproject.dao.EmployeeDAO;
 import com.example.springboot.restproject.entity.Employee;
 import com.example.springboot.restproject.service.EmployeeService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api")
 public class EmployeeRestController {
 
+    private ObjectMapper objectMapper;
     private EmployeeService employeeService;
 
     @Autowired
-    public EmployeeRestController(EmployeeService employeeService){
+    public EmployeeRestController(EmployeeService employeeService,ObjectMapper objectMapper){
         this.employeeService = employeeService;
+        this.objectMapper = objectMapper;
     }
 
     @GetMapping("/employees")
@@ -45,8 +49,47 @@ public class EmployeeRestController {
         return employeeInDB;
     }
 
+
+
+    @PatchMapping("/employees/{employeeId}")
+    public Employee patchEmployee(@PathVariable int employeeId,@RequestBody Map<String,Object> patchPayload){
+        Employee employeeInDB = employeeService.findById(employeeId);
+
+        if(employeeInDB == null){
+            throw new RuntimeException("Employee id not found - "+employeeId);
+        }
+
+        if(patchPayload.containsKey("id")){
+            throw  new RuntimeException("Employee id not allowed in request body - "+employeeId);
+        }
+
+        Employee patchedEmployee = apply(employeeInDB,patchPayload);
+        Employee updateEmployeeInDB = employeeService.save(patchedEmployee);
+
+        return patchedEmployee;
+
+    }
+
     @DeleteMapping("/employees/{employeeId}")
-    public void removeEmployee(@PathVariable int employeeId){
+    public String removeEmployee(@PathVariable int employeeId){
+        Employee employeeInDB = employeeService.findById(employeeId);
+        if(employeeInDB == null){
+            throw new RuntimeException("Employee id not found - "+employeeId);
+        }
         employeeService.deleteById(employeeId);
+
+        return "Deleted Employee with id - "+employeeId;
+    }
+
+    public Employee apply(Employee employee,Map<String,Object> patchPayload){
+        ObjectNode employeeNode = objectMapper.convertValue(employee, ObjectNode.class);
+        ObjectNode patchPayloadNode = objectMapper.convertValue(patchPayload, ObjectNode.class);
+
+        System.out.println(employeeNode);
+        System.out.println(patchPayloadNode);
+
+        employeeNode.setAll(patchPayloadNode);
+
+        return  objectMapper.convertValue(employeeNode,Employee.class);
     }
 }
